fix: disabled sitemap
