+++ app.config.ts
@@ -4,7 +4,6 @@ import tailwindcss from "@tailwindcss/vite";
 import { defineConfig } from "@tanstack/react-start/config";
 import unfonts from "unplugin-fonts/vite";
 import tsConfigPaths from "vite-tsconfig-paths";
-import generateSitemap, { sitemap } from "./src/plugins/sitemap";
 
 const config = defineConfig({
   vite: {
@@ -32,7 +31,6 @@ const config = defineConfig({
           families: ["Geist", "Geist Mono"],
         },
       }),
-      generateSitemap(sitemap),
     ],
     ssr: {
       noExternal: ["react-use", "react-markdown"],
