+++ package.json
@@ -111,6 +111,7 @@
     "husky": "^9.1.7",
     "lint-staged": "^15.5.2",
     "tailwindcss": "^4.1.5",
+    "tanstack-router-sitemap": "^1.0.12",
     "typescript": "^5.8.3",
     "vite-tsconfig-paths": "^5.1.4"
   },
