+++ tsconfig.json
@@ -15,8 +15,7 @@
     "forceConsistentCasingInFileNames": true,
     "baseUrl": ".",
     "paths": {
-      "~/*": ["./src/*"],
-      "@/*": ["./*"]
+      "~/*": ["./src/*"]
     },
     "noEmit": true,
     "strictNullChecks": true,
